directory-upload

+++ frontend/src/components/Sidebar.tsx
@@ -1,7 +1,8 @@
 import { useNavigate } from 'react-router-dom';
-import { Pencil, Trash, Upload } from 'lucide-react';
+import { Upload } from 'lucide-react';
 import UploadFileButton from './UploadFileButton';
 import { useState } from "react";
+import FileTree, { buildTree, NodeType } from "@/components/FileTree.tsx";
 
 type SidebarProps = {
   files: { key : string }[];
@@ -13,15 +14,20 @@ type SidebarProps = {
 const Sidebar = ({ files = [], onFileSelect, onFileDelete, onRefreshFiles } : SidebarProps) => {
   const navigate = useNavigate();
 
-  const [editingFileKey, seteditingFileKey] = useState<string | null>(null);
+  const [editingFileKey, setEditingFileKey] = useState<string | null>(null);
+  const [editingFileType, setEditingFileType] = useState<NodeType | null>(null);
   const [newFileKey, setNewFileKey] = useState("");
 
-  const handleDelete = async(fileKey : string) => {
+  const handleDelete = async(fileKey : string, type: NodeType | null) => {
+    if ( !type ) {
+      // TODO: Error handling.
+      return;
+    }
     if(!window.confirm('Are you sure you want to permanently remove the file?')) return;
 
     onFileDelete(fileKey);
     try {
-      const response = await fetch(`http://localhost:5001/delete/${fileKey}`, {
+      const response = await fetch(`http://localhost:5001/` + (type == "file" ? "delete/" : "delete-dir/") + fileKey, {
         method: 'DELETE'
       });
 
@@ -35,15 +41,20 @@ const Sidebar = ({ files = [], onFileSelect, onFileDelete, onRefreshFiles } : Si
     }
   };
 
-  const handleRename = async (fileKey : string, newFileKey: string) => {
+  const handleRename = (fileKey: string) => async (newFileKey: string, type: NodeType | null) => {
+    if ( !type ) {
+      // TODO: Error handling.
+      return;
+    }
     try {
-      const response = await fetch(`http://localhost:5001/rename/${fileKey}`, {
+      const response = await fetch(`http://localhost:5001/` + (type == "file" ? "rename/" : "rename-dir/") + fileKey, {
         method: 'PATCH',
         headers: {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
-          content: newFileKey
+          // TODO: This logic should really all be in backend.
+          content: type == "file" ? fileKey.replace(fileKey.split("/").pop() ?? "", newFileKey) : newFileKey,
         })
       });
 
@@ -72,38 +83,12 @@ const Sidebar = ({ files = [], onFileSelect, onFileDelete, onRefreshFiles } : Si
       </div>
 
       <div className="flex-1">
-        {files.map((file, index) => (
-          <div
-            key={index}
-            className="group flex items-center justify-between -4 py-2 rounded-lg hover:bg-white/10 cursor-pointer transition-colors"
-          >
-            <div
-              className = "cursor-pointer flex-1 truncate"
-              onClick={() => onFileSelect(file.key)} // pass selected file key up
-            >
-              {file.key}
-            </div>
-            <div className = "flex items-center gap-2 ml-2 opacity-0 group-hover:opacity-100 transition-opacity duration-200">
-              <Pencil
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    seteditingFileKey(file.key);
-                    setNewFileKey(file.key);
-                  }}
-                  className="size-6 p-1 rounded-md hover:bg-gray-200 cursor-pointer"
-              />
-            </div>
-            <div className = "flex items-center gap-2 ml-2 opacity-0 group-hover:opacity-100 transition-opacity duration-200"> 
-              <Trash 
-              onClick={(e) => {
-                e.stopPropagation();
-                handleDelete(file.key)
-              }}
-                className="size-6 p-1 rounded-md hover:bg-gray-200 curser-pointer"
-              />
-            </div>
-          </div>
-        ))}
+        <FileTree treeData={buildTree(files.map(obj => obj.key))} onDelete={handleDelete} onSelect={onFileSelect} onEdit={(name : string, type: NodeType) => {
+            setEditingFileKey(name);
+            setNewFileKey(name);
+            setEditingFileType(type);
+          }
+        } />
       </div>
 
       <div className="mt-4">
@@ -111,7 +96,7 @@ const Sidebar = ({ files = [], onFileSelect, onFileDelete, onRefreshFiles } : Si
           <Upload className="mx-auto mb-2" />
           <p>Drop files here</p>
           <p className="text-sm text-gray-400 mt-2">Or</p>
-          <div className="mt-2 px-4 py-2 bg-blue-600 rounded-lg hover:bg-blue-700 transition-colors">
+          <div style={{ display: "flex", flexDirection: "column", gap: "5px" }} >
             <UploadFileButton onUploadComplete={onRefreshFiles} />
           </div>
         </div>
@@ -120,24 +105,24 @@ const Sidebar = ({ files = [], onFileSelect, onFileDelete, onRefreshFiles } : Si
   {editingFileKey && (
       <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
         <div className="bg-white p-6 rounded-lg shadow-lg w-96">
-          <h2 className="text-lg font-bold mb-4">Rename File</h2>
+          <h2 className="text-lg font-bold mb-4">Rename {editingFileType}</h2>
           <input
               type="text"
-              value={newFileKey}
+              value={newFileKey.split("/").pop() || ""}
               onChange={(e) => setNewFileKey(e.target.value)}
               className="w-full px-3 py-2 border rounded mb-4"
           />
           <div className="flex justify-end gap-2">
             <button
-                onClick={() => seteditingFileKey(null)}
+                onClick={() => setEditingFileKey(null)}
                 className="px-4 py-2 bg-gray-200 rounded hover:bg-gray-300"
             >
               Cancel
             </button>
             <button
                 onClick={async () => {
-                  await handleRename(editingFileKey, newFileKey);
-                  seteditingFileKey(null);
+                  await handleRename(editingFileKey)(newFileKey, editingFileType);
+                  setEditingFileKey(null);
                 }}
                 className="px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700"
             >
